modify folder structure and add try catch for user routes and controller
